Salesforce: start splitting into sync vs async
