+++ airbyte-integrations/connectors/source-salesforce/source_salesforce/async_salesforce/__init__.py
@@ -1,0 +1,3 @@
+from .source import AsyncSourceSalesforce, SalesforceSourceDispatcher
+
+__all__ = ["AsyncSourceSalesforce", "SalesforceSourceDispatcher"]
