+++ airbyte-integrations/connectors/source-salesforce/source_salesforce/async_salesforce/availability_strategy.py
@@ -8,15 +8,16 @@ from typing import Optional, Tuple
 
 from airbyte_cdk.sources.async_cdk.streams.http.availability_strategy_async import AsyncHttpAvailabilityStrategy
 from airbyte_cdk.sources.streams import Stream
-from requests import HTTPError, codes
+from airbyte_cdk.sources.streams.http.utils import HttpError
+from requests import codes
 
 if typing.TYPE_CHECKING:
     from airbyte_cdk.sources import Source
 
 
-class SalesforceAvailabilityStrategy(AsyncHttpAvailabilityStrategy):
+class AsyncSalesforceAvailabilityStrategy(AsyncHttpAvailabilityStrategy):
     def handle_http_error(
-        self, stream: Stream, logger: logging.Logger, source: Optional["Source"], error: HTTPError
+        self, stream: Stream, logger: logging.Logger, source: Optional["Source"], error: HttpError
     ) -> Tuple[bool, Optional[str]]:
         """
         There are several types of Salesforce sobjects that require additional processing:
@@ -26,8 +27,8 @@ class SalesforceAvailabilityStrategy(AsyncHttpAvailabilityStrategy):
              And since we use a dynamic method of generating streams for Salesforce connector - at the stage of discover,
              we cannot filter out these streams, so we check for them before reading from the streams.
         """
-        if error.response.status_code in [codes.FORBIDDEN, codes.BAD_REQUEST]:
-            error_data = error.response.json()[0]
+        if error.status_code in [codes.FORBIDDEN, codes.BAD_REQUEST]:
+            error_data = error.json()[0]
             error_code = error_data.get("errorCode", "")
             if error_code != "REQUEST_LIMIT_EXCEEDED" or error_code == "INVALID_TYPE_FOR_OPERATION":
                 return False, f"Cannot receive data for stream '{stream.name}', error message: '{error_data.get('message')}'"
