+++ airbyte-integrations/connectors/source-salesforce/source_salesforce/async_salesforce/exceptions.py
